Revert "Cast time range properties to int instead of long"
This reverts commit f0039cbe8160d10c7b21ecc2d98840272e69f84a.

+++ src/main/java/com/google/sps/utilities/RealSearchDatastore.java
@@ -133,8 +133,8 @@ public final class RealSearchDatastore implements SearchDatastoreService {
     * @return TimeRange
     */
     private TimeRange createTimeRange(Entity entity) {
-        int start = (int) entity.getProperty("start");
-        int end = (int) entity.getProperty("end");
+        int start = Math.toIntExact((long) entity.getProperty("start"));
+        int end = Math.toIntExact((long) entity.getProperty("end"));
         Calendar date = new Gson().fromJson((String) entity.getProperty("date"), Calendar.class);
 
         return TimeRange.fromStartToEnd(start, end, date);
